+++ storage/store.go
@@ -39,7 +39,7 @@ type Store interface {
 	ReadLastMintDistribution(group string) (*common.MintDistribution, error)
 	LockMintInput(mint *common.MintData, tx crypto.Hash, fork bool) error
 	ReadMintDistributions(group string, offset, count uint64) ([]*common.MintDistribution, []*common.VersionedTransaction, error)
-	ListNodeWorks(cids []crypto.Hash, day uint64) (map[crypto.Hash][2]uint64, error)
+	ListNodeWorks(cids []crypto.Hash, day uint32) (map[crypto.Hash][2]uint64, error)
 	ReadWorkOffset(nodeId crypto.Hash) (uint64, error)
 	WriteRoundWork(nodeId crypto.Hash, round uint64, snapshots []*common.SnapshotWithTopologicalOrder) error
